Rewrite autoloader so we have all rules defined in separate class
The class with defined rules now is reusable and we can use it
also on rewriting the libraries

We need to think what we can do with functions,
and somehow I think it will be better to create these alias
in repositories where we have defined functions.

We would need define list of all possible function defined in
libraries and then just call new function using `call_user_func_array`
and `func_get_args`.