fix: better detection of Stratigility functions
The functions only exist when we have a `CallableMiddlewareDecorator`
class, so this patch now tests for that class being present before
attempting to define the functions.

Additionally, it checks to see if the PSR-15 middleware compatibility
shim interface is present. If so, it defines only the middleware,
doublePass, and path functions, and does so using the appropriate
typehints. If the PSR-15 interfaces are present, it will instead define
the functions defined in Stratigility 3.0.