Adds typehints compatibility
User code and 3rd party libraries still use legacy typehints.
If we try to load new class and corresponding legacy class was not
loaded before then we can get error of typehints incompatibility.
On loading new class we need then create alias to legacy class
so PHP interpreter thinks that it's the same class and we can
have seamless migration.