feat: config post processing
One common issue when using third-party libraries is that the
configuration keys will still reference ZF artifacts: namespaces,
classes, and configuration keys.

This patch provides a recursive configuration post-processor that
rewrites those references using the rules used when performing
migrations, ensuring that they now reference the Laminas equivalents.

The post processor, `Laminas\ZendFrameworkBridge\ConfigPostProcessor`
can be used with laminas-config-aggregator. Additionally, you can
register `Laminas\ZendFrameworkBridge` as a module in MVC applications,
and it will register a `ModuleEvent::EVENT_CONFIG_MERGE` listener that
will pass merged configuration to the post processor, and pass the
transformed configuration back to the `ConfigListener`.

The module is NOT exposed via the `extra.laminas.module` directive, to
ensure that it is opt-in, and thus can also be disabled.

The migration tooling will be updated to register the module in MVC
applications, and the post processor in Expressive applications.

I also recommend that we have the migrations package depend on this one,
so that all rewrite rules are in one package.