fix: ensure we rewrite alias _targets_
This patch is actually a refactor of how replacements are made to use a
rules engine. Each rule is run, and can return null or a callable. If a
null is returned, it checks the next rule, and so-forth. If a callable
is returned, it is used to perform the replacement.

This approach allows us to have rules of different granularity, and for
specific conditions.

I rewrote the "exact replacements" functionality as a ruleset, and added
two more:

- one for when the key is "aliases" and it has an array value
- one for a non-null key and an array value

In the first case, it now invokes the new method
`replaceDependencyAliases`, which replaces the targets for aliases, but
not the keys. In the second case, it calls `__invoke()` on the value.

The fallback checks for a string value, returning the result of
`Replacements::replace()` on it if so, and the unchanged value
otherwise.

The `replaceValue()` method was renamed `replace`, and is now called
both for keys (in which case the `$key` value passed to a ruleset is
`null`) as well as values.