feat: add support for processing invokables
This patch does two things.

First, it modifies how replacements are matched and triggered, passing a
hierarchical array of keys so that rulesets can determine where they are
in the hierarchy, as well as access the keys present. This allows us to,
for instance, know when we are matching a top-level key.

Second, it adds a ruleset for rewriting invokables configuration. If the
alias does not change but the target does, then that change is made. If
the alias changes, too, then:

- An entry is added for the original alias, pointing to the rewritten target.
- An entry is added for the new alias, pointing to the rewritten target.

This ensures aliases work as expected, and the original service names
are still present in configuration.