fix: better rules for invokables
Invokables are essentially an alias map. This patch modifies the code to
remove the separate logic for rewriting invokables, and instead have the
rewrite rule for aliases also handle invokables.

That way, the rule becomes: rewrite the target only. This ensures the
original service name persists, but now maps to the new target class.