bugfix: allow the class loader to be absent
With composer 2.2.0, some autoloaders are loaded even before the `vendor/autoload.php` file is dumped.
This leads to installation crashes in some circumstances due to the fact that this component throws a `RuntimeException` in case the autoloader could not be found.

Due to the fact, that this works properly after the autoloader got dumped, we can safely not register the prepend/append autoloader registration for this case.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>