feat: allow configuring the ConfigPostProcessor
Per a suggestion in #94, this patch modifies the `ConfigPostProcessor` such that it now examines the merged configuration passed to it for its own configuration, specifically for replacement rules.
These may be provided via the following configuration:

```php
return [
    'laminas-zendframework-bridge' => [
        'replacements' => [
            'to-replace' => 'replacement',
            // ...
        ],
    ],
];
```

This configuration itself will NEVER be rewritten.
However, the rules will now be used as additional replacements when running the processor.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>